Move LinalgQuantized* passes to GlobalOptimization
These passes are a better fit for GlobalOptimization given they have to
do with Linalg ops. Additionally this turns on the passes for all
frontends now that other frontends are starting to generate these ops
too.
