+++ compiler/src/iree/compiler/GlobalOptimization/BUILD.bazel
@@ -61,6 +61,8 @@ iree_compiler_cc_library(
         "OptimizeNumerics.cpp",
         "Passes.cpp",
         "PropagateLinalgTranspose.cpp",
+        "QuantizedConvToConv.cpp",
+        "QuantizedMatmulToMatmul.cpp",
         "RaiseSpecialOps.cpp",
         "RemoveZeroExtentTensors.cpp",
         "SimplifyPackUnpack.cpp",
