+++ compiler/src/iree/compiler/GlobalOptimization/CMakeLists.txt
@@ -57,6 +57,8 @@ iree_cc_library(
     "OptimizeNumerics.cpp"
     "Passes.cpp"
     "PropagateLinalgTranspose.cpp"
+    "QuantizedConvToConv.cpp"
+    "QuantizedMatmulToMatmul.cpp"
     "RaiseSpecialOps.cpp"
     "RemoveZeroExtentTensors.cpp"
     "SimplifyPackUnpack.cpp"
