+++ compiler/src/iree/compiler/InputConversion/Common/test/CMakeLists.txt
@@ -19,8 +19,6 @@ iree_lit_test_suite(
     "demote_i64_to_i32.mlir"
     "import_ml_program.mlir"
     "iree_import_public.mlir"
-    "linalg_quantized_conv_to_conv.mlir"
-    "linalg_quantized_matmul_to_matmul.mlir"
     "promote_bf16_to_f32.mlir"
     "promote_f16_to_f32.mlir"
     "sanitize_module_names.mlir"
